requirements.txt
# Core
aiohttp>=3.9.0
cryptography>=41.0.0
websockets>=12.0
click>=8.1.0
rich>=13.7.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Numerics / ML
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
joblib>=1.3.0

# Storage
aiosqlite>=0.19.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.0
python-dateutil>=2.8.2
pytz>=2023.3
colorama>=0.4.6
tabulate>=0.9.0
prompt-toolkit>=3.0.43
